fix(addon-mobile): fix minLength and maxLength properties for range in mobile calendar
