+++ projects/kit/components/calendar-range/calendar-range.component.ts
@@ -21,7 +21,7 @@ import {
 import {tuiWatch} from '@taiga-ui/cdk/observables';
 import {TuiMapperPipe} from '@taiga-ui/cdk/pipes/mapper';
 import type {TuiBooleanHandler, TuiMapper} from '@taiga-ui/cdk/types';
-import {tuiIsString, tuiNullableSame, tuiPure} from '@taiga-ui/cdk/utils/miscellaneous';
+import {tuiIsString, tuiNullableSame} from '@taiga-ui/cdk/utils/miscellaneous';
 import type {TuiMarkerHandler} from '@taiga-ui/core/components/calendar';
 import {TuiCalendar} from '@taiga-ui/core/components/calendar';
 import {TuiDataList} from '@taiga-ui/core/components/data-list';
@@ -30,6 +30,7 @@ import {TUI_COMMON_ICONS} from '@taiga-ui/core/tokens';
 import {TUI_CALENDAR_DATE_STREAM, TUI_OTHER_DATE_TEXT} from '@taiga-ui/kit/tokens';
 import type {Observable} from 'rxjs';
 
+import {calculateDisabledItemHandler} from './calculate-disabled-item-handler';
 import {TUI_DAY_CAPS_MAPPER} from './day-caps-mapper';
 import type {TuiDayRangePeriod} from './day-range-period';
 
@@ -123,7 +124,7 @@ export class TuiCalendarRange implements OnInit, OnChanges {
     }
 
     protected get calculatedDisabledItemHandler(): TuiBooleanHandler<TuiDay> {
-        return this.calculateDisabledItemHandler(
+        return calculateDisabledItemHandler(
             this.disabledItemHandler,
             this.value,
             this.minLength,
@@ -227,29 +228,6 @@ export class TuiCalendarRange implements OnInit, OnChanges {
             ) ||
                 null)
         );
-    }
-
-    @tuiPure
-    private calculateDisabledItemHandler(
-        disabledItemHandler: TuiBooleanHandler<TuiDay>,
-        value: TuiDayRange | null,
-        minLength: TuiDayLike | null,
-    ): TuiBooleanHandler<TuiDay> {
-        return (item) => {
-            if (!value?.isSingleDay || !minLength) {
-                return disabledItemHandler(item);
-            }
-
-            const negativeMinLength = Object.fromEntries(
-                Object.entries(minLength).map(([key, value]) => [key, -value]),
-            );
-            const disabledBefore = value.from.append(negativeMinLength).append({day: 1});
-            const disabledAfter = value.from.append(minLength).append({day: -1});
-            const inDisabledRange =
-                disabledBefore.dayBefore(item) && disabledAfter.dayAfter(item);
-
-            return inDisabledRange || disabledItemHandler(item);
-        };
     }
 
     private initDefaultViewedMonth(): void {
