+++ projects/kit/components/calendar-range/index.ts
@@ -1,3 +1,4 @@
+export * from './calculate-disabled-item-handler';
 export * from './calendar-range.component';
 export * from './day-caps-mapper';
 export * from './day-range-period';
